backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
pydantic
python-multipart
passlib[bcrypt]
pyjwt
aiofiles
Pillow
numpy
matplotlib
librosa
soundfile
pytest
opencv-python
scikit-image

